refactor app.py (7/7): extract bunny helpers to helpers/bunny.py
Five helpers move to helpers/bunny.py:
- list_bunny_ad_folder
- optimize_image
- upload_status_image_to_bunny
- _allowed_file
- _upload_to_bunny

All Bunny env-var constants (BUNNY_*, BUNNY_STATUS_*, BUNNY_AD_*) and
ALLOWED_FILE_EXTENSIONS are redefined locally in helpers/bunny.py from
os.environ — same fallback chain as app.py, production unaffected.

list_bunny_ad_folder's app.logger.error call swapped to a standard
logging.getLogger(__name__) so the helper doesn't import app.

Smoke-tested: /publish, /command-deck/, /cockpit/after-dark/library
(the last calls list_bunny_ad_folder which returns [] locally without
Bunny AD env vars set — but the route returns 200 cleanly).

Final sub-phase. Helpers extraction complete: 20 functions across
7 helper files, app.py shrunk by ~250 lines of helper definitions.
Ready for end-to-end verification before PR.
